fix(review): move marketplace install test to test_tools_dispatch_validation.py
test_dispatch_food_truck_marketplace_install_succeeds tests execution
behavior with a MarketplaceInstall plugin source, not parameter passthrough
(resume, idle_timeout). Moving to the validation file preserves naming
symmetry for the _params module.
